completed recipe routes and components

+++ app/forms/recipe_ingredients_form.py
@@ -1,8 +1,8 @@
-from wtforms import IntegerField, Form
+from wtforms import IntegerField, Form, FloatField
 from wtforms.validators import DataRequired
 
 
 class RecipeIngredientForm(Form):
     ingredient_id = IntegerField('ingredientId', validators=[DataRequired()])
     measurement_id = IntegerField('measurementId', validators=[DataRequired()])
-    quantity = IntegerField('quantity', validators=[DataRequired()])
+    quantity = FloatField('quantity', validators=[DataRequired()])
